feat: update application configuration and enhance DTOs for natural language queries
