+++ pom.xml
@@ -220,9 +220,9 @@
                 <artifactId>flyway-maven-plugin</artifactId>
                 <version>${flyway.version}</version>
                 <configuration>
-                    <url>jdbc:mysql://localhost:3306/insight_data</url>
-                    <user>root</user>
-                    <password>root</password>
+                    <url>jdbc:mysql://localhost:3306/datainsight</url>
+                    <user>test</user>
+                    <password>test</password>
                     <locations>
                         <location>classpath:db/migration</location>
                     </locations>
