+++ src/main/java/com/insightdata/facade/rest/dto/SavedQueryDTO.java
@@ -1,4 +1,4 @@
-package com.insightdata.api.dto;
+package com.insightdata.facade.rest.dto;
 
 import java.util.Map;
 
@@ -8,13 +8,18 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * 自然语言查询请求DTO
+ * 保存查询请求DTO
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class NLQueryRequestDTO {
+public class SavedQueryDTO {
+    
+    /**
+     * 查询名称
+     */
+    private String name;
     
     /**
      * 数据源ID
@@ -27,7 +32,22 @@ public class NLQueryRequestDTO {
     private String query;
     
     /**
+     * 生成的SQL
+     */
+    private String sql;
+    
+    /**
      * 查询参数
      */
     private Map<String, Object> parameters;
+    
+    /**
+     * 描述
+     */
+    private String description;
+    
+    /**
+     * 是否公开
+     */
+    private boolean isPublic;
 }
