+++ src/main/resources/application-dev.yml
@@ -1,9 +1,9 @@
 spring:
   # 数据库配置
   datasource:
-    url: jdbc:mysql://localhost:3306/insight_data?useUnicode=true&characterEncoding=utf8&useSSL=false&serverTimezone=Asia/Shanghai
-    username: root
-    password: root
+    url: jdbc:mysql://localhost:3306/datainsight?useUnicode=true&characterEncoding=utf8&useSSL=false&serverTimezone=Asia/Shanghai
+    username: test
+    password: test
   
   # JPA配置
   jpa:
